Keep CLI file length within guard after merge
The merge from origin/main moved CLI/cmux.swift over the Swift file-length budget. The install-preview renderer was already a self-contained CMUXCLI static helper, so it now lives in CMUXCLI+InstallPreview.swift and is registered with the CLI target without changing behavior.

Constraint: Do not edit .yml or .yaml files while updating this PR.

Rejected: Increase the Swift file-length budget | the guard failure is better handled by splitting a self-contained helper out of the oversized file.

Confidence: high

Scope-risk: narrow

Directive: Keep future cmux.swift growth behind extension files when helpers are self-contained.

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv; swiftc -parse CLI/cmux.swift CLI/CMUXCLI+InstallPreview.swift; plutil -lint GhosttyTabs.xcodeproj/project.pbxproj; git diff --check

Not-tested: Local app build/tests per repo policy; PR CI will validate after push.
